refactor: move meta.relative_wealth_index to metrics
