--- tests/test_population_weighted_relative_wealth-index.py
@@ -1,51 +1,0 @@
-"""Tests for functions in population-weighted/relative-wealth-index.py."""
-
-from unittest import mock
-
-import geopandas as gpd
-import pandas as pd
-
-from dart_pipeline.population_weighted.relative_wealth_index import (
-    process_gadm_popdensity_rwi,
-)
-
-
-@mock.patch("geopandas.read_file")
-@mock.patch("pandas.read_csv")
-def test_process_gadm_popdensity_rwi(mock_pd_read, mock_gpd_read):
-    # Mock GADM shapefile
-    mock_gpd_read.return_value = gpd.GeoDataFrame(
-        pd.DataFrame(
-            {
-                "GID_0": ["GID0"],
-                "COUNTRY": ["Mockland"],
-                "NAME_1": ["NAME1"],
-                "NAME_2": ["NAME2"],
-                "GID_2": ["GID2"],
-            }
-        ),
-        geometry=gpd.points_from_xy([106.0], [10.0]),
-        crs="EPSG:4326",
-    )
-
-    # Mock CSV imports
-    mock_pd_read.side_effect = [
-        # Mock RWI data
-        pd.DataFrame({"latitude": [10.0], "longitude": [106.0], "rwi": [0.5]}),
-        # Mock population data
-        pd.DataFrame(
-            {
-                "latitude": [10.0],
-                "longitude": [106.0],
-                "quadkey": ["1234"],
-                "pop_2023": [1000],
-            }
-        ),
-    ]
-
-    df = process_gadm_popdensity_rwi("VNM", "2023", "2", plots=False)
-
-    assert isinstance(df, pd.DataFrame)
-    assert "iso3" in df.columns
-    assert "value" in df.columns
-    assert not df.empty
